document compile contract
The internal documentaion is a little lighter than
usual in `src/compile.rs`, so I wrote up the impression
of the compilation contract I developed by reading the
code. This is honestly a thinly veiled way to check my
own understanding of what is going on. Hopefully it will
be valuable to the next person who starts digging into
the compiler.