ci: limit testing on 1.12.0
The latest update to rand requires a newer version of Rust. Since it's
a dev dependency, we shouldn't need to do a semver bump when updating
rand. However, CI needs to be told not to run tests. Instead, we merely
check that we can build the crate and produce documentation.