literal: fix anchor performance problem
The Match literal iterator would repeatedly look for matches
in the remainder of the input after it found its first
match regardless of whether or not the regex was anchored
at the start. This commit adds logic to make sure that we
don't keep looking for matches after the first match is
returned for a start-anchored literal regex.