literal: auto enable SIMD on Rust stable 1.27+
This commit removes the need to use the `unstable` feature to enable
SIMD optimizations. We add a "version sniffer" to the `build.rs` script
to detect if Rust version 1.27 or newer is being used, and if so, enable
the SIMD optimizations.

The 'unstable' feature is now a no-op, but we keep it for backwards
compatibility. We also may use it again some day.