input: fix infinite loop regression
This fixes a bug introduced by a bug fix for #557. In particular, the
termination condition wasn't exactly right, and this appears to have
slipped through the test suite. This probably reveals a hole in our test
suite, which is specifically the testing of Unicode regexes with
bytes::Regex on invalid UTF-8.

This bug was originally reported against ripgrep:
https://github.com/BurntSushi/ripgrep/issues/1247