teddy: make u8x16 and u8x32 have vector call ABI
Before this commit, u8x16 and u8x32 were repr(Rust) unions. This
introduced unspecified behavior because the field offsets of repr(Rust)
unions are not guaranteed to be at offset 0, so that field access was
potentially UB. This commit fixes that.

The unions were also generating a lot of unnecessary memory operations.
This commit fixes that as well.

The issue is that unions have an Aggregate call ABI, which is the same
as the call ABI of arrays. That is, they are passed around by memory,
and not in Vector registers.

This is good, if most of the time one operates on them as arrays. This
was, however, not the case. Most of the operations on these unions are
using SIMD instructions. This means that the union needs to be copied
into a SIMD register, operated on, and then spilled back to the stack,
on every single operation. That's unnecessary, although apparently LLVM
was able to optimize all the unnecessary memory operations away and
leave these always in registers.

This commit fixes this issue as well, by making the u8x16 and u8x32
repr(transparent) newtypes over the architecture specific vector types,
giving them the Vector ABI.

The vectors are then copied to the stack only when necessary, and as
little as possible. This is done using mem::transmute, removing the
need for unions altogether (fixing #588 by not having to worry about
union layout at all).

To make it clear when the vectors are spilled into the stack, the
vector::replace(index, value) and vector::extract(index) APIs
have been removed, and instead, only a vector::bytes(self) and a
vector::from_bytes(&mut self, [u8; N]) APIs are provided instead. This
prevents spilling the vectors back and forth onto the stack every time
an index needs to be modified, by using vector::bytes to spill the
vector to the stack once, making all the random-access modifications in
memory, and then using vector::from_bytes only once to move the memory
back into a SIMD register.

Closes #588, Closes #589