literal: remove teddy
The Teddy algorithm has been moved to the aho-corasick crate (as of
0.7.5), so we can now use it there.

Note that we do explicitly use aho-corasick's `packed` module instead of
relying on `AhoCorasick`'s prefilter to do it for us. The reasoning,
unfortunately, is that using Teddy inside of `AhoCorasick` has some
measurable overhead that we'd like to avoid. It would be better to
figure out how to remove that overhead, but I was unsuccessful. It isn't
much additional work to reach around and used the packed search
directly.

Benchmarks roughly stay the  same, with the `regexdna::variant`
benchmarks improving across the board by about 1.5x-2x. Some benchmarks
do regress (e.g., `sherlock::the_nocase`), but we decide to live with it
for now. The regression is likely due to subtle changes in Teddy's
bucket allocation.